+++ commands.go
@@ -10,7 +10,6 @@ package main
 import (
 	"errors"
 	"fmt"
-	"github.com/kennygrant/sanitize"
 	"github.com/layeh/gumble/gumble"
 	"os"
 	"regexp"
@@ -24,10 +23,11 @@ import (
 // it contains a command.
 func parseCommand(user *gumble.User, username, command string) {
 	var com, argument string
-	sanitizedCommand := sanitize.HTML(command)
-	if strings.Contains(sanitizedCommand, " ") {
-		index := strings.Index(sanitizedCommand, " ")
-		com, argument = sanitizedCommand[0:index], sanitizedCommand[(index+1):]
+	split := strings.Split(command, "\n")
+	splitString := split[0]
+	if strings.Contains(splitString, " ") {
+		index := strings.Index(splitString, " ")
+		com, argument = splitString[0:index], splitString[(index+1):]
 	} else {
 		com = command
 		argument = ""
@@ -39,35 +39,35 @@ func parseCommand(user *gumble.User, username, command string) {
 		if dj.HasPermission(username, dj.conf.Permissions.AdminAdd) {
 			add(user, username, argument)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Skip command
 	case dj.conf.Aliases.SkipAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminSkip) {
 			skip(user, username, false, false)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Skip playlist command
 	case dj.conf.Aliases.SkipPlaylistAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminAddPlaylists) {
 			skip(user, username, false, true)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Forceskip command
 	case dj.conf.Aliases.AdminSkipAlias:
 		if dj.HasPermission(username, true) {
 			skip(user, username, true, false)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Playlist forceskip command
 	case dj.conf.Aliases.AdminSkipPlaylistAlias:
 		if dj.HasPermission(username, true) {
 			skip(user, username, true, true)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Stop command
 	case dj.conf.Aliases.StopAlias:
@@ -103,66 +103,73 @@ func parseCommand(user *gumble.User, username, command string) {
 		if dj.HasPermission(username, dj.conf.Permissions.AdminHelp) {
 			help(user)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Volume command
 	case dj.conf.Aliases.VolumeAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminVolume) {
 			volume(user, username, argument)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Move command
 	case dj.conf.Aliases.MoveAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminMove) {
 			move(user, argument)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Reload command
 	case dj.conf.Aliases.ReloadAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminReload) {
 			reload(user)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Reset command
 	case dj.conf.Aliases.ResetAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminReset) {
 			reset(username)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Numsongs command
 	case dj.conf.Aliases.NumSongsAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminNumSongs) {
 			numSongs()
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Nextsong command
 	case dj.conf.Aliases.NextSongAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminNextSong) {
 			nextSong(user)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Currentsong command
 	case dj.conf.Aliases.CurrentSongAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminCurrentSong) {
 			currentSong(user)
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
+		}
+	// Setcomment command
+	case dj.conf.Aliases.SetCommentAlias:
+		if dj.HasPermission(username, dj.conf.Permissions.AdminSetComment) {
+			setComment(user, argument)
+		} else {
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	// Kill command
 	case dj.conf.Aliases.KillAlias:
 		if dj.HasPermission(username, dj.conf.Permissions.AdminKill) {
 			kill()
 		} else {
-			user.Send(NO_PERMISSION_MSG)
+			dj.SendPrivateMessage(user, NO_PERMISSION_MSG)
 		}
 	default:
-		user.Send(COMMAND_DOESNT_EXIST_MSG)
+		dj.SendPrivateMessage(user, COMMAND_DOESNT_EXIST_MSG)
 	}
 }
 
@@ -170,7 +177,7 @@ func parseCommand(user *gumble.User, username, command string) {
 // the URL to the queue if the format matches.
 func add(user *gumble.User, username, url string) {
 	if url == "" {
-		user.Send(NO_ARGUMENT_MSG)
+		dj.SendPrivateMessage(user, NO_ARGUMENT_MSG)
 	} else {
 		youtubePatterns := []string{
 			`https?:\/\/www\.youtube\.com\/watch\?v=([\w-]+)`,
@@ -193,17 +200,20 @@ func add(user *gumble.User, username, url string) {
 		}
 
 		if matchFound {
-			newSong := NewSong(username, shortUrl)
-			if err := dj.queue.AddItem(newSong); err == nil {
-				dj.client.Self().Channel().Send(fmt.Sprintf(SONG_ADDED_HTML, username, newSong.title), false)
-				if dj.queue.Len() == 1 && !dj.audioStream.IsPlaying() {
-					if err := dj.queue.CurrentItem().(*Song).Download(); err == nil {
-						dj.queue.CurrentItem().(*Song).Play()
-					} else {
-						user.Send(AUDIO_FAIL_MSG)
-						dj.queue.CurrentItem().(*Song).Delete()
+			if newSong, err := NewSong(username, shortUrl, nil); err == nil {
+				if err := dj.queue.AddSong(newSong); err == nil {
+					dj.client.Self.Channel.Send(fmt.Sprintf(SONG_ADDED_HTML, username, newSong.title), false)
+					if dj.queue.Len() == 1 && !dj.audioStream.IsPlaying() {
+						if err := dj.queue.CurrentSong().Download(); err == nil {
+							dj.queue.CurrentSong().Play()
+						} else {
+							dj.SendPrivateMessage(user, AUDIO_FAIL_MSG)
+							dj.queue.CurrentSong().Delete()
+						}
 					}
 				}
+			} else {
+				dj.SendPrivateMessage(user, INVALID_YOUTUBE_ID_MSG)
 			}
 		} else {
 			// Check to see if we have a playlist URL instead.
@@ -212,23 +222,25 @@ func add(user *gumble.User, username, url string) {
 				if re.MatchString(url) {
 					if dj.HasPermission(username, dj.conf.Permissions.AdminAddPlaylists) {
 						shortUrl = re.FindStringSubmatch(url)[1]
-						newPlaylist := NewPlaylist(username, shortUrl)
-						if dj.queue.AddItem(newPlaylist); err == nil {
-							dj.client.Self().Channel().Send(fmt.Sprintf(PLAYLIST_ADDED_HTML, username, newPlaylist.title), false)
-							if dj.queue.Len() == 1 && !dj.audioStream.IsPlaying() {
-								if err := dj.queue.CurrentItem().(*Playlist).songs.CurrentItem().(*Song).Download(); err == nil {
-									dj.queue.CurrentItem().(*Playlist).songs.CurrentItem().(*Song).Play()
+						oldLength := dj.queue.Len()
+						if newPlaylist, err := NewPlaylist(username, shortUrl); err == nil {
+							dj.client.Self.Channel.Send(fmt.Sprintf(PLAYLIST_ADDED_HTML, username, newPlaylist.title), false)
+							if oldLength == 0 && !dj.audioStream.IsPlaying() {
+								if err := dj.queue.CurrentSong().Download(); err == nil {
+									dj.queue.CurrentSong().Play()
 								} else {
-									user.Send(AUDIO_FAIL_MSG)
-									dj.queue.CurrentItem().(*Playlist).songs.CurrentItem().(*Song).Delete()
+									dj.SendPrivateMessage(user, AUDIO_FAIL_MSG)
+									dj.queue.CurrentSong().Delete()
 								}
 							}
+						} else {
+							dj.SendPrivateMessage(user, INVALID_YOUTUBE_ID_MSG)
 						}
 					} else {
-						user.Send(NO_PLAYLIST_PERMISSION_MSG)
+						dj.SendPrivateMessage(user, NO_PLAYLIST_PERMISSION_MSG)
 					}
 				} else {
-					user.Send(INVALID_URL_MSG)
+					dj.SendPrivateMessage(user, INVALID_URL_MSG)
 				}
 			}
 		}
@@ -240,39 +252,46 @@ func add(user *gumble.User, username, url string) {
 func skip(user *gumble.User, username string, admin, playlistSkip bool) {
 	if dj.audioStream.IsPlaying() {
 		if playlistSkip {
-			if dj.queue.CurrentItem().ItemType() == "playlist" {
-				if err := dj.queue.CurrentItem().AddSkip(username); err == nil {
+			if dj.queue.CurrentSong().playlist != nil {
+				if err := dj.queue.CurrentSong().playlist.AddSkip(username); err == nil {
 					if admin {
-						dj.client.Self().Channel().Send(ADMIN_PLAYLIST_SKIP_MSG, false)
+						dj.client.Self.Channel.Send(ADMIN_PLAYLIST_SKIP_MSG, false)
 					} else {
-						dj.client.Self().Channel().Send(fmt.Sprintf(PLAYLIST_SKIP_ADDED_HTML, username), false)
+						dj.client.Self.Channel.Send(fmt.Sprintf(PLAYLIST_SKIP_ADDED_HTML, username), false)
 					}
-					if dj.queue.CurrentItem().SkipReached(len(dj.client.Self().Channel().Users())) || admin {
-						dj.queue.CurrentItem().(*Playlist).skipped = true
-						dj.client.Self().Channel().Send(PLAYLIST_SKIPPED_HTML, false)
+					if dj.queue.CurrentSong().playlist.SkipReached(len(dj.client.Self.Channel.Users)) || admin {
+						id := dj.queue.CurrentSong().playlist.id
+						dj.queue.CurrentSong().playlist.DeleteSkippers()
+						for i := 0; i < len(dj.queue.queue); i++ {
+							if dj.queue.queue[i].playlist != nil {
+								if dj.queue.queue[i].playlist.id == id {
+									dj.queue.queue = append(dj.queue.queue[:i], dj.queue.queue[i+1:]...)
+									i--
+								}
+							}
+						}
+						if dj.queue.Len() != 0 {
+							// Set dontSkip to true to avoid audioStream.Stop() callback skipping the new first song.
+							dj.queue.CurrentSong().dontSkip = true
+						}
+						dj.client.Self.Channel.Send(PLAYLIST_SKIPPED_HTML, false)
 						if err := dj.audioStream.Stop(); err != nil {
 							panic(errors.New("An error occurred while stopping the current song."))
 						}
 					}
 				}
 			} else {
-				user.Send(NO_PLAYLIST_PLAYING_MSG)
+				dj.SendPrivateMessage(user, NO_PLAYLIST_PLAYING_MSG)
 			}
 		} else {
-			var currentItem QueueItem
-			if dj.queue.CurrentItem().ItemType() == "playlist" {
-				currentItem = dj.queue.CurrentItem().(*Playlist).songs.CurrentItem()
-			} else {
-				currentItem = dj.queue.CurrentItem()
-			}
-			if err := currentItem.AddSkip(username); err == nil {
+			if err := dj.queue.CurrentSong().AddSkip(username); err == nil {
 				if admin {
-					dj.client.Self().Channel().Send(ADMIN_SONG_SKIP_MSG, false)
+					dj.client.Self.Channel.Send(ADMIN_SONG_SKIP_MSG, false)
 				} else {
-					dj.client.Self().Channel().Send(fmt.Sprintf(SKIP_ADDED_HTML, username), false)
+					dj.client.Self.Channel.Send(fmt.Sprintf(SKIP_ADDED_HTML, username), false)
 				}
-				if currentItem.SkipReached(len(dj.client.Self().Channel().Users())) || admin {
-					dj.client.Self().Channel().Send(SONG_SKIPPED_HTML, false)
+				if dj.queue.CurrentSong().SkipReached(len(dj.client.Self.Channel.Users)) || admin {
+					dj.client.Self.Channel.Send(SONG_SKIPPED_HTML, false)
 					if err := dj.audioStream.Stop(); err != nil {
 						panic(errors.New("An error occurred while stopping the current song."))
 					}
@@ -280,7 +299,7 @@ func skip(user *gumble.User, username string, admin, playlistSkip bool) {
 			}
 		}
 	} else {
-		user.Send(NO_MUSIC_PLAYING_MSG)
+		dj.SendPrivateMessage(user, NO_MUSIC_PLAYING_MSG)
 	}
 }
 
@@ -309,7 +328,7 @@ func SoundBoard(username string, argument string, lerandom bool) {
 
 // Performs help functionality. Displays a list of valid commands.
 func help(user *gumble.User) {
-	user.Send(fmt.Sprintf(HELP_HTML, dj.conf.SoundBoardList.SoundBoardSite))
+	dj.SendPrivateMessage(user, HELP_HTML)
 }
 
 // Performs volume functionality. Checks input value against LowestVolume and HighestVolume from
@@ -317,18 +336,18 @@ func help(user *gumble.User) {
 // is applied and is immediately in effect.
 func volume(user *gumble.User, username, value string) {
 	if value == "" {
-		dj.client.Self().Channel().Send(fmt.Sprintf(CUR_VOLUME_HTML, dj.audioStream.Volume()), false)
+		dj.client.Self.Channel.Send(fmt.Sprintf(CUR_VOLUME_HTML, dj.audioStream.Volume), false)
 	} else {
 		if parsedVolume, err := strconv.ParseFloat(value, 32); err == nil {
 			newVolume := float32(parsedVolume)
 			if newVolume >= dj.conf.Volume.LowestVolume && newVolume <= dj.conf.Volume.HighestVolume {
-				dj.audioStream.SetVolume(newVolume)
-				dj.client.Self().Channel().Send(fmt.Sprintf(VOLUME_SUCCESS_HTML, username, dj.audioStream.Volume()), false)
+				dj.audioStream.Volume = newVolume
+				dj.client.Self.Channel.Send(fmt.Sprintf(VOLUME_SUCCESS_HTML, username, dj.audioStream.Volume), false)
 			} else {
-				user.Send(fmt.Sprintf(NOT_IN_VOLUME_RANGE_MSG, dj.conf.Volume.LowestVolume, dj.conf.Volume.HighestVolume))
+				dj.SendPrivateMessage(user, fmt.Sprintf(NOT_IN_VOLUME_RANGE_MSG, dj.conf.Volume.LowestVolume, dj.conf.Volume.HighestVolume))
 			}
 		} else {
-			user.Send(fmt.Sprintf(NOT_IN_VOLUME_RANGE_MSG, dj.conf.Volume.LowestVolume, dj.conf.Volume.HighestVolume))
+			dj.SendPrivateMessage(user, fmt.Sprintf(NOT_IN_VOLUME_RANGE_MSG, dj.conf.Volume.LowestVolume, dj.conf.Volume.HighestVolume))
 		}
 	}
 }
@@ -337,12 +356,12 @@ func volume(user *gumble.User, username, value string) {
 // to the channel if it is.
 func move(user *gumble.User, channel string) {
 	if channel == "" {
-		user.Send(NO_ARGUMENT_MSG)
+		dj.SendPrivateMessage(user, NO_ARGUMENT_MSG)
 	} else {
-		if dj.client.Channels().Find(channel) != nil {
-			dj.client.Self().Move(dj.client.Channels().Find(channel))
+		if dj.client.Channels.Find(channel) != nil {
+			dj.client.Self.Move(dj.client.Channels.Find(channel))
 		} else {
-			user.Send(CHANNEL_DOES_NOT_EXIST_MSG)
+			dj.SendPrivateMessage(user, CHANNEL_DOES_NOT_EXIST_MSG)
 		}
 	}
 }
@@ -350,9 +369,7 @@ func move(user *gumble.User, channel string) {
 // Performs reload functionality. Tells command submitter if the reload completed successfully.
 func reload(user *gumble.User) {
 	if err := loadConfiguration(); err == nil {
-		user.Send(CONFIG_RELOAD_SUCCESS_MSG)
-	} else {
-		panic(err)
+		dj.SendPrivateMessage(user, CONFIG_RELOAD_SUCCESS_MSG)
 	}
 }
 
@@ -360,12 +377,13 @@ func reload(user *gumble.User) {
 // remaining songs in the ~/.mumbledj/songs directory.
 func reset(username string) {
 	dj.queue.queue = dj.queue.queue[:0]
-	if err := dj.audioStream.Stop(); err == nil {
-		if err := deleteSongs(); err == nil {
-			dj.client.Self().Channel().Send(fmt.Sprintf(QUEUE_RESET_HTML, username), false)
-		} else {
+	if dj.audioStream.IsPlaying() {
+		if err := dj.audioStream.Stop(); err != nil {
 			panic(err)
 		}
+	}
+	if err := deleteSongs(); err == nil {
+		dj.client.Self.Channel.Send(fmt.Sprintf(QUEUE_RESET_HTML, username), false)
 	} else {
 		panic(err)
 	}
@@ -376,10 +394,10 @@ func reset(username string) {
 // the number of songs in the queue to chat.
 func numSongs() {
 	songCount := 0
-	dj.queue.Traverse(func(i int, item QueueItem) {
+	dj.queue.Traverse(func(i int, song *Song) {
 		songCount += 1
 	})
-	dj.client.Self().Channel().Send(fmt.Sprintf(NUM_SONGS_HTML, songCount), false)
+	dj.client.Self.Channel.Send(fmt.Sprintf(NUM_SONGS_HTML, songCount), false)
 }
 
 // Performs nextsong functionality. Uses the SongQueue PeekNext function to peek at the next
@@ -387,9 +405,9 @@ func numSongs() {
 // of the next item if it exists.
 func nextSong(user *gumble.User) {
 	if song, err := dj.queue.PeekNext(); err != nil {
-		user.Send(NO_SONG_NEXT_MSG)
+		dj.SendPrivateMessage(user, NO_SONG_NEXT_MSG)
 	} else {
-		user.Send(fmt.Sprintf(NEXT_SONG_HTML, song.title, song.submitter))
+		dj.SendPrivateMessage(user, fmt.Sprintf(NEXT_SONG_HTML, song.title, song.submitter))
 	}
 }
 
@@ -397,16 +415,21 @@ func nextSong(user *gumble.User) {
 // information about the song currently playing.
 func currentSong(user *gumble.User) {
 	if dj.audioStream.IsPlaying() {
-		var currentItem *Song
-		if dj.queue.CurrentItem().ItemType() == "playlist" {
-			currentItem = dj.queue.CurrentItem().(*Playlist).songs.CurrentItem().(*Song)
+		if dj.queue.CurrentSong().playlist == nil {
+			dj.SendPrivateMessage(user, fmt.Sprintf(CURRENT_SONG_HTML, dj.queue.CurrentSong().title, dj.queue.CurrentSong().submitter))
 		} else {
-			currentItem = dj.queue.CurrentItem().(*Song)
+			dj.SendPrivateMessage(user, fmt.Sprintf(CURRENT_SONG_PLAYLIST_HTML, dj.queue.CurrentSong().title,
+				dj.queue.CurrentSong().submitter, dj.queue.CurrentSong().playlist.title))
 		}
-		user.Send(fmt.Sprintf(CURRENT_SONG_HTML, currentItem.title, currentItem.submitter))
 	} else {
-		user.Send(NO_MUSIC_PLAYING_MSG)
+		dj.SendPrivateMessage(user, NO_MUSIC_PLAYING_MSG)
 	}
+}
+
+// Performs setcomment functionality. Sets the bot's comment to whatever text is supplied in the argument.
+func setComment(user *gumble.User, comment string) {
+	dj.client.Self.SetComment(comment)
+	dj.SendPrivateMessage(user, COMMENT_UPDATED_MSG)
 }
 
 // Performs kill functionality. First cleans the ~/.mumbledj/songs directory to get rid of any