+++ strings.go
@@ -48,7 +48,13 @@ const ADMIN_SONG_SKIP_MSG = "An admin has decided to skip the current song."
 const ADMIN_PLAYLIST_SKIP_MSG = "An admin has decided to skip the current playlist."
 
 // Message shown to users when the audio for a video could not be downloaded.
-const AUDIO_FAIL_MSG = "The audio download for this video failed. YouTube has likely not generated the audio files for this video yet."
+const AUDIO_FAIL_MSG = "The audio download for this video failed. YouTube has likely not generated the audio files for this video yet. Skipping to the next song!"
+
+// Message shown to users when they supply a YouTube URL that does not contain a valid ID.
+const INVALID_YOUTUBE_ID_MSG = "The YouTube URL you supplied did not contain a valid YouTube ID."
+
+// Message shown to user when they successfully update the bot's comment.
+const COMMENT_UPDATED_MSG = "The comment for the bot has successfully been updated."
 
 // Message shown to a channel when a new song starts playing.
 const NOW_PLAYING_HTML = `
@@ -61,6 +67,24 @@ const NOW_PLAYING_HTML = `
 		</tr>
 		<tr>
 			<td align="center">Added by %s</td>
+		</tr>
+	</table>
+`
+
+// Message shown to channel when a new song in a playlist starts playing.
+const NOW_PLAYING_PLAYLIST_HTML = `
+	<table>
+		<tr>
+			<td align="center"><img src="%s" width=150 /></td>
+		</tr>
+		<tr>
+			<td align="center"><b><a href="http://youtu.be/%s">%s</a> (%s)</b></td>
+		</tr>
+		<tr>
+			<td align="center">Added by %s</td>
+		</tr>
+		<tr>
+			<td align="center">From playlist "%s"</td>
 		</tr>
 	</table>
 `
@@ -106,6 +130,7 @@ const HELP_HTML = `<br/>
 	<p><b>!forceskipplaylist</b> - An admin command that forces a playlist skip. </p>
 	<p><b>!move </b>- Moves MumbleDJ into channel if it exists.</p>
 	<p><b>!reload</b> - Reloads mumbledj.gcfg configuration settings.</p>
+	<p><b>!setcomment</b> - Sets the comment for the bot.</p>
 	<p><b>!kill</b> - Safely cleans the bot environment and disconnects from the server.</p>
 `
 
@@ -147,4 +172,10 @@ const NEXT_SONG_HTML = `
 // Message shown to users when they issue the currentsong command.
 const CURRENT_SONG_HTML = `
 	The song currently playing is "%s", added by <b>%s</b>.
+`
+
+// Message shown to users when the currentsong command is issued when a song from a
+// playlist is playing.
+const CURRENT_SONG_PLAYLIST_HTML = `
+	The song currently playing is "%s", added <b>%s</b> from the playlist "%s".
 `